fix(reservation-domain): TestFixture 에서 assertj 의존성 제거
